+++ lib/main.dart
@@ -2,8 +2,7 @@ import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:pocket_pose/data/local/provider/video_play_provider.dart';
-import 'package:pocket_pose/data/local/provider/auth_provider.dart';
-import 'package:pocket_pose/data/remote/provider/signin_signup_provider.dart';
+import 'package:pocket_pose/data/remote/provider/auth_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:pocket_pose/ui/screen/main_screen.dart';
 
@@ -20,7 +19,6 @@ Future<void> main() async {
   runApp(MultiProvider(providers: [
     ChangeNotifierProvider(create: (_) => VideoPlayProvider()),
     ChangeNotifierProvider(create: (_) => AuthProvider()),
-    ChangeNotifierProvider(create: (_) => SignInSignUpProvider()),
   ], child: const MyApp()));
 }
